requirements.txt
cohere==5.3.2
gradio==4.27.0
langchain==0.1.16
langchain_community==0.0.34
langchain_openai==0.1.3
pandas==2.2.2
python-dotenv==1.0.1